moved Client class from hyped::client::Client to hyped::telemetry::Client (like it should have been from the beginning)
