+++ src/telemetry/signalhandler.cpp
@@ -26,9 +26,9 @@ namespace telemetry {
 
 bool SignalHandler::receivedSigPipeSignal = false;
 
-// didn't really want to set this as a static variable and hardcode logger values 
+// didn't really want to set this as a static variable and hardcode logger values
 // but didn't really see any other way (sys.verbose_tlm gets set at runtime)
-Logger SignalHandler::log_(true, 2);  
+Logger SignalHandler::log_(true, 2);
 
 SignalHandler::SignalHandler()
 {
